// backend/src/lista-compras/dto/retorno-lista-compra.dto.ts
export class ProdutoListaDTO {
  idProduto: number;
  codProduto: string;
  idCategoria: number;
  nomeProduto: string;
  preco: number;
  promocao: boolean;
  quantidadeAtual: number;
  quantidadeComprada: number;
  quantidadeVendida: number;
  statusProduto: boolean;
  descricao: string;
  dimensoes: string;
  precoPromocao: number;
  categoria: string;
}